combine test files
